+++ packages/core/client/src/modules/blocks/filter-blocks/collapse/createFilterCollapseBlockSchema.ts
